Move `useDeepMemo` to new utils directory
We'll keep the Hooks in the root of the project as the ones that
we're exporting.
